Build openMHA for Debian 11 and Ubuntu 22.04
Summary:
Adjust Jenkinsfiles and make necessary adjustments to the packaging Makefile and to test_windnoise.m to cope with incompatible changes encountered on the new platforms.

Ubuntu LTS 22.04 which will be released next month and we want to provide openMHA for it, therefore we must start building for it.

The Debian version 10 that we currently use as the basis for the Mahalia image will stop receiving security updates in August, therefore we should also start supporting the successor Debian 11 in order to be able to migrate Mahalia.

The build environments incl. the non-standard build dependencies (liblsl, mhamakedeb) have already been published on apt.hoertech.de for Ubuntu 22.04 ("Jammy") and Debian 11 ("Bullseye"), see the history documented in T1813.

In addition to supporting the new targets, this change removes the ARM64 Bionic target. Reason: People who use openMHA on ARM64 are very likely to use more modern systems than Ubuntu 18.04. Our last release 1.17.0 for that target will not go away, but we do not need to compile the next release for Ubuntu 18.04 ARM64 when we can compile for Debian 11 instead, the binaries for which will also work on Ubuntu 22.04.

Test Plan:
A Test build on Jenkins with Replay is available here https://jenkins.hz-ol.de/blue/organizations/jenkins/MHA%2Fcode_reviews/detail/code_reviews/259/pipeline and was successful. Previous build attempts show:
* https://jenkins.hz-ol.de/blue/organizations/jenkins/MHA%2Fcode_reviews/detail/code_reviews/256/pipeline
** A problem with Octave on ARM-Debian-11 that has been worked around by changing test_windnoise.m.
* The Change in the Makefile is explained here: T1813#33637  (search for WARNING: Make 4.3 Backward-incompatibility!)

Reviewers: giso, hendrikk, pmaanen

Reviewed By: giso

Maniphest Tasks: T1813

Differential Revision: https://dev.openmha.org/D695