Update installation instructions to cover new target systems
Summary:
Ubuntu 22.04 has been released, and we compile for it.  Update
installation instructions to cover this system.

Debian 10, on which Mahalia is based, will reach its end-of-life
in August this year.  We may switch from Debian 10 to Debian 11
as the base for Mahalia, and from openMHA compiled for Bionic on
ARM to openMHA compiled for Bullseye on ARM.  Prepare by
providing updated installation instructions for ARM.

Installation instructions follow recommendations on the man page
of apt key (see e.g.
https://manpages.ubuntu.com/manpages/jammy/man8/apt-key.8.html,
section "DEPRECATION").

Test Plan:
I have tested the installation instructions for the additional
systems on

1) Ubuntu 22.04 on x86_64
2) Debian 10 on ARM32
3) Debian 10 on ARM64
4) Ubuntu 22.04 on ARM32
5) Ubuntu 22.04 on ARM64

Differential Revision: https://dev.openmha.org/D708