Replace calls to sprintf by calls to snprintf.
Summary:
The former is deprecated for security reasons and even though the warning
does not really apply to us because we control the input to sprintf, it's
less work to just please the compiler than to change the warning settings.

Test Plan:
Manual review. Make sure that the input sizes to snprintf are equal to
the sizes of the buffers used. Grep through the code to make sure no
instance is missed.

Reviewers: tobias

Reviewed By: tobias

Maniphest Tasks: T1843

Differential Revision: https://dev.openmha.org/D721